MB-10533 Move `format_failure_msg_from_exception` to rest.py
Fits better in here
